music_cards_service/requirements.txt
asgiref==3.4.1
boto3==1.18.23
botocore==1.21.23
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorthief==0.2.1
Django==3.2.6
httplib2==0.19.1
idna==3.2
image==1.5.33
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
snakeviz==2.1.0
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
